+++ Stajs.BalloonicornHunter.Core/Server/Player.cs
@@ -10,7 +10,7 @@ namespace Stajs.BalloonicornHunter.Core.Server
 	{
 		public int Index { get; set; }
 		public string Name { get; set; }
-		public long Score { get; set; }
+		public int Score { get; set; }
 		public float Duration { get; set; }
 	}
 }
